moved parse_args to separate file
