+++ src/__init__.py
@@ -24,3 +24,4 @@ from .matching import (
     quantile_filter,
     left_median_filter,
 )
+from .parse_args import parse_args
